% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGeneration.R
\name{PAGeneration}
\alias{PAGeneration}
\title{PAGeneration}
\usage{
PAGeneration(data, nb_pa, ratio_pa_in)
}
\arguments{
\item{data}{list of length 2 created by SpeciesNiche (1 = an n-dimensions array, 2 = name of each dimensions (environmental variables))}

\item{nb_pa}{a vector of number of pseudo-absences you want to generate}

\item{ratio_pa_in}{a vector of ratio of pseudo-absences in and out the species niche (for details and advice see help file at : )}
}
\value{
a list of data-frame of length nb_pa*ratio_pa_InOut with environmental values of PA in n columns + a column saying if the PA come from the inside or the outside of the species niche + a column saying the weight of sampling of the PA. Those two last columns allow the user to verify that the PA generated match with its demand.
}
\description{
PAGeneration
}
\examples{
set.seed(10)
data_species = data.frame(Env_Var1= rnorm(100),
                              Env_Var2 = rnorm(100,3,1),
                              Env_Var3 = rnorm(100,5,2))

species_niche = SpeciesNiche(data = data_species, hist_type = c("FD", "FD", "FD"), niche_border = c(-3,3, 0,6, 0,10))

PAGeneration(data = species_niche, nb_pa = c(1000, 3000, 5000), ratio_pa_in = c(1, 2/3))


}
