% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesNiche.R
\name{SpeciesNiche}
\alias{SpeciesNiche}
\title{SpeciesNiche}
\usage{
SpeciesNiche(data, niche_border = NULL, bins_sizes)
}
\arguments{
\item{data}{a dataframe with m rows = presences of the species and n columns = environmental variables}

\item{bins_sizes}{wich formula is going to be used to calculate bins sizes for each environmental variables, based on the 3 types of the base hist funtion in r : "Sturges"; "Scott"; "FD" (Freedman-Diaconis)}

\item{niche_border}{by default (if not specified) the minimun and maximum of each environmental variables in the data; if specified must be : min of env var 1, max of env var 1, min of env var 2, max of env var 2, ..., min of env var n, max of env var n}
}
\value{
an n-dimensions array, each dimensions being an environmental variables, filled with the number of presences for each associations of those variables
}
\description{
SpeciesNiche
}
\examples{
set.seed(10)
data_species = data.frame(Pred_1= rnorm(100),
                              Pred_2 = rnorm(100,3,1),
                              Pred_3 = rnorm(100,5,2))

SpeciesNiche(data = data_species, niche_border = c(-3,3, 0,6, 0,10), bins_sizes = c("FD", "FD", "FD"))

}
