% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesNiche.R
\name{SpeciesNiche}
\alias{SpeciesNiche}
\title{SpeciesNiche
This function construct an ecological space, each of its dimensions being one of the environmental predictors composing "data".
This ecological space is an array. An array need two informations to be constructed (1) the minimum and maximum values of each dimensions in "niche_border" and (2) each dimensions being sliced into bins regularly space, the size of those bins in "bins_sizes".
Once the array is constructed it is filled with the number of presences observed for each associations of environmental values.}
\usage{
SpeciesNiche(data, niche_border = NULL, bins_sizes)
}
\arguments{
\item{data}{a dataframe with m rows = presences of the species and n columns = environmental variables}

\item{niche_border}{by default (if not specified) the minimum and maximum of each environmental variables in the data; if specified must be : min of env var 1, max of env var 1, min of env var 2, max of env var 2, ..., min of env var n, max of env var n}

\item{bins_sizes}{a vector of length number of predictors (column of data) specifying the size of the bins constructing each dimension of the array. Can be numeric values decided by the user or algorithms calculating the optimal size of the bin ("Sturges", "Freedman-Diaconis", "Scott", see details in hist base r function). Numeric values and algorithm can not be mixed.}
}
\value{
a list of length 3 with (1) an n-dimensions array, each dimensions being an environmental variables, filled with the number of presences for each associations of those variables, (2) the names of each dimensions (environmental variables) and (3) the input presences of "data" round to the size of the bins. We strongly recommend to use the round presences (+ pseudo-absences generated in PAGeneration) in your Species Distribution Modeling procedure.
}
\description{
SpeciesNiche
This function construct an ecological space, each of its dimensions being one of the environmental predictors composing "data".
This ecological space is an array. An array need two informations to be constructed (1) the minimum and maximum values of each dimensions in "niche_border" and (2) each dimensions being sliced into bins regularly space, the size of those bins in "bins_sizes".
Once the array is constructed it is filled with the number of presences observed for each associations of environmental values.
}
\examples{
set.seed(10)
data_species = data.frame(Pred_1= rnorm(100),
                          Pred_2 = rnorm(100,3,1),
                          Pred_3 = rnorm(100,5,2))

SpeciesNiche(data = data_species, niche_border = c(-3,3, 0,6, 0,10), bins_sizes = c("FD", "FD", "FD"))

}
